backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.121.0",
    "passlib>=1.7.4",
    "pydantic[email]>=2.12.3",
    "pydantic-settings>=2.11.0",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "alembic>=1.17.1",
    "redis>=7.0.1",
    "fastapi-cache2>=0.2.2",
    "aiobotocore>=2.25.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.4",
]

test = [
    "pytest>=9.0.0",
    "httpx>=0.28.1",
    "asgi-lifespan>=2.1.0",
    "greenlet>=3.0.3",
    "pytest-asyncio>=1.3.0",
]

[tool.ruff]
target-version = "py311"
exclude = ["venv", ".venv", ".env", "src/typings/external"]

line-length = 120
lint.select = ["ALL"]
lint.ignore = [
    # modules
    "C90",        # mccabe complexity
    "DJ",         # django
    "T10",        # debugger

    # specific rules
    "COM",        # flake8-commas
    "S104",       # binging to all interfaces
    "D100",       # ignore missing docs
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D200",       # ignore one-line docstring
    "D203",
    "D212",
    "D213",
    "D400",
    "D415",
    "TRY003",     # external messages in exceptions are too verbose
    "TD002",
    "TD003",
    "FIX002",     # too verbose descriptions of todos
    "S101",       # prohibited assert statement
    "ARG001",     # unused function argument
    "ANN401",     # disallow Any type
    "RUF001",     # string contrains ambigious o
    "RUF002",     # other ambigious cyrillic symbols in string
    "RUF003",     # yet another ambigious cyrillic symbol warning
    "PLR0913",    # too many arguments in function signature
]
format.skip-magic-trailing-comma = false

[tool.ruff.lint.per-file-ignores]
"alembic/*" = ["INP001"]
"tests/*" = [
    "ANN",       # Type annotations
    "INP001",    # Missing __init__.py
    "S105",      # Possible hardcoded password
    "S106",      # Possible hardcoded password assigned to argument  
]
